third_party/blink/web_tests/external/wpt/app-history/currentchange-event/event-constructor.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
  assert_throws_js(TypeError, () => {
    new AppHistoryCurrentChangeEvent("currententrychange");
  });
}, "can't bypass required members by omitting the dictionary entirely");

test(() => {
  assert_throws_js(TypeError, () => {
    new AppHistoryCurrentChangeEvent("currententrychange", {
      navigationType: "push"
    });
  });
}, "from is required");

test(() => {
  const event = new AppHistoryCurrentChangeEvent("currententrychange", {
    navigationType: "replace",
    from: navigation.currentEntry
  });
  assert_equals(event.navigationType, "replace");
  assert_equals(event.from, navigation.currentEntry);
}, "all properties are reflected back");

test(t => {
  const event = new AppHistoryCurrentChangeEvent("currententrychange", { from: navigation.currentEntry });
  assert_equals(event.navigationType, null);
}, "defaults are as expected");
</script>
